--- time/src/main/java/net/jqwik/time/api/constraints/DayOfMonth.java
@@ -1,22 +1,0 @@
-package net.jqwik.time.api.constraints;
-
-import java.lang.annotation.*;
-
-import org.apiguardian.api.*;
-
-import static org.apiguardian.api.API.Status.*;
-
-/**
- * Constrain Integers to be valid days of month.
- * <p>
- * Applies to int or Integer parameters which are also annotated with {@code @ForAll}.
- *
- * @see net.jqwik.api.ForAll
- */
-@Target({ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@API(status = EXPERIMENTAL, since = "1.4.1")
-public @interface DayOfMonth {
-
-}
