+++ time/src/main/java/net/jqwik/time/internal/properties/configurators/DayOfMonthRangeConfigurator.java
@@ -13,9 +13,7 @@ public class DayOfMonthRangeConfigurator extends ArbitraryConfiguratorBase {
 
 	@Override
 	protected boolean acceptTargetType(TypeUsage targetType) {
-		return targetType.isAssignableFrom(LocalDate.class) || (targetType.isAssignableFrom(Integer.class) && targetType
-																													  .findAnnotation(DayOfMonth.class)
-																													  .isPresent());
+		return targetType.isAssignableFrom(LocalDate.class) || targetType.isAssignableFrom(Integer.class);
 	}
 
 	public Arbitrary<?> configure(Arbitrary<?> arbitrary, DayOfMonthRange range) {
