+++ time/src/main/java/net/jqwik/time/internal/properties/providers/MonthDayArbitraryProvider.java
@@ -1,4 +1,4 @@
-package net.jqwik.time.providers;
+package net.jqwik.time.internal.properties.providers;
 
 import java.time.*;
 import java.util.*;
@@ -7,14 +7,14 @@ import net.jqwik.api.*;
 import net.jqwik.api.providers.*;
 import net.jqwik.time.api.*;
 
-public class YearArbitraryProvider implements ArbitraryProvider {
+public class MonthDayArbitraryProvider implements ArbitraryProvider {
 	@Override
 	public boolean canProvideFor(TypeUsage targetType) {
-		return targetType.isAssignableFrom(Year.class);
+		return targetType.isAssignableFrom(MonthDay.class);
 	}
 
 	@Override
 	public Set<Arbitrary<?>> provideFor(TypeUsage targetType, SubtypeProvider subtypeProvider) {
-		return Collections.singleton(Dates.years());
+		return Collections.singleton(Dates.monthDays());
 	}
 }
