+++ time/src/main/resources/META-INF/services/net.jqwik.api.providers.ArbitraryProvider
@@ -1,5 +1,4 @@
-net.jqwik.time.providers.DatesArbitraryProvider
-net.jqwik.time.providers.DayOfMonthArbitraryProvider
-net.jqwik.time.providers.MonthDayArbitraryProvider
-net.jqwik.time.providers.YearArbitraryProvider
-net.jqwik.time.providers.YearMonthArbitraryProvider
+net.jqwik.time.internal.properties.providers.DatesArbitraryProvider
+net.jqwik.time.internal.properties.providers.MonthDayArbitraryProvider
+net.jqwik.time.internal.properties.providers.YearArbitraryProvider
+net.jqwik.time.internal.properties.providers.YearMonthArbitraryProvider
